chore(asm): refactor: move code from module to iast
